Consolidate getCookies function
