--- src/utilities/utils.ts
@@ -1,7 +1,0 @@
-
-// getCookie: used to obtain the given name stored in browser cookies
-export function getCookie(name) {
-    var re = new RegExp(name + "=([^;]+)");
-    var value = re.exec(document.cookie);
-    return (value != null) ? decodeURI(value[1]) : null;
-}
