+++ src/utils/cookies.tsx
@@ -1,7 +1,7 @@
 // Takes a reference to document.cookie and the name of the desired cookie
-function getCookie(docCookies, cookieName) {
-    var re = new RegExp(cookieName + "=([^;]+)");
-    var value = re.exec(docCookies);
+function getCookie(name) {
+    var re = new RegExp(name + "=([^;]+)");
+    var value = re.exec(document.cookie);
     return (value != null) ? decodeURI(value[1]) : null;
 }
